No more multiplatform for core
