+++ core/common/src/main/kotlin/com/boswelja/smartwatchextensions/core/devicemanagement/VersionSerializer.kt
