+++ core/common/src/main/kotlin/com/boswelja/smartwatchextensions/core/settings/BoolSettingSerializer.kt
