+++ core/common/src/main/kotlin/com/boswelja/smartwatchextensions/core/settings/IntSettingSerializer.kt
