+++ core/mobile/build.gradle.kts
@@ -2,46 +2,39 @@
 @file:Suppress("UnstableApiUsage", "DSL_SCOPE_VIOLATION")
 
 plugins {
-    id("com.boswelja.smartwatchextensions.library")
+    kotlin("android")
+    id("com.android.library")
     id("com.boswelja.smartwatchextensions.detekt")
     id("com.squareup.sqldelight")
     kotlin("plugin.serialization") version "1.6.10"
     alias(libs.plugins.compose)
 }
 
-kotlin {
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(projects.core.common)
-                implementation(libs.watchconnection.mobile.core)
-                implementation(libs.sqldelight.runtime)
-                implementation(libs.sqldelight.coroutines)
-                implementation(libs.koin.core)
-                implementation(libs.kotlinx.serialization.protobuf)
-            }
-        }
-        val commonTest by getting {
-            dependencies {
-                implementation(kotlin("test"))
-                implementation(libs.turbine)
-                implementation(libs.kotlinx.coroutines.test)
-            }
-        }
-        val androidMain by getting {
-            dependencies {
-                implementation(libs.sqldelight.android)
-                implementation(libs.bundles.compose.mobile)
-                implementation(libs.androidx.datastore.proto)
-                implementation(libs.koin.android)
-            }
-        }
-        val androidTest by getting {
-            dependencies {
-                implementation(libs.sqldelight.sqlitedriver)
-            }
-        }
+android {
+    compileSdk = 32
+    defaultConfig {
+        minSdk = 26
+        targetSdk = 32
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
+}
+
+dependencies {
+    api(projects.core.common)
+    implementation(libs.watchconnection.mobile.core)
+    implementation(libs.sqldelight.runtime)
+    implementation(libs.sqldelight.coroutines)
+    implementation(libs.koin.core)
+    implementation(libs.kotlinx.serialization.protobuf)
+    implementation(libs.sqldelight.android)
+    implementation(libs.bundles.compose.mobile)
+    implementation(libs.androidx.datastore.proto)
+    implementation(libs.koin.android)
+
+    testImplementation("junit:junit:4.13.2")
+    testImplementation(libs.turbine)
+    testImplementation(libs.kotlinx.coroutines.test)
+    testImplementation(libs.sqldelight.sqlitedriver)
 }
 
 sqldelight {
@@ -52,5 +45,11 @@ sqldelight {
     database("RegisteredWatchDatabase") {
         packageName = "com.boswelja.smartwatchextensions.core.devicemanagement.database"
         sourceFolders = listOf("database-watches")
+    }
+}
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
+    kotlinOptions {
+        freeCompilerArgs = freeCompilerArgs + "-Xopt-in=kotlin.RequiresOptIn"
     }
 }
