--- core/mobile/src/androidTest/kotlin/com/boswelja/smartwatchextensions/core/settings/WatchSettingsDbRepositoryTest.kt
@@ -1,6 +1,0 @@
-package com.boswelja.smartwatchextensions.core.settings
-
-import com.squareup.sqldelight.db.SqlDriver
-import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver
-
-actual fun createTestSqlDriver(): SqlDriver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
