+++ core/mobile/src/main/AndroidManifest.xml
