+++ core/mobile/src/main/database-settings/com/boswelja/smartwatchextensions/core/settings/database/BoolSetting.sq
