+++ core/mobile/src/main/database-watches/com/boswelja/smartwatchextensions/core/devicemanagement/database/RegisteredWatch.sq
