+++ core/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/core/DI.kt
