+++ core/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/core/UpdateChecker.kt
