+++ core/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/core/devicemanagement/SelectedWatchState.kt
