+++ core/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/core/settings/WatchSettingsDbRepository.kt
