+++ core/mobile/src/main/kotlin/com/boswelja/smartwatchextensions/core/ui/settings/CheckboxSetting.kt
