+++ core/mobile/src/main/res/values/strings.xml
