+++ core/wear/build.gradle.kts
@@ -1,22 +1,28 @@
 plugins {
-    id("com.boswelja.smartwatchextensions.library")
+    kotlin("android")
+    id("com.android.library")
     id("com.boswelja.smartwatchextensions.detekt")
     kotlin("plugin.serialization") version "1.6.10"
 }
 
-kotlin {
-    sourceSets {
-        val commonMain by getting {
-            dependencies {
-                api(projects.core.common)
-                implementation(libs.kotlinx.serialization.protobuf)
-            }
-        }
-        val androidMain by getting {
-            dependencies {
-                implementation(libs.androidx.datastore.proto)
-                implementation(libs.koin.android)
-            }
-        }
+android {
+    compileSdk = 32
+    defaultConfig {
+        minSdk = 26
+        targetSdk = 32
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+}
+
+dependencies {
+    api(projects.core.common)
+    implementation(libs.kotlinx.serialization.protobuf)
+    implementation(libs.androidx.datastore.proto)
+    implementation(libs.koin.android)
+}
+
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
+    kotlinOptions {
+        freeCompilerArgs = freeCompilerArgs + "-Xopt-in=kotlin.RequiresOptIn"
     }
 }
