+++ core/wear/src/main/AndroidManifest.xml
